--- src/pages/Login.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Login = () => {
-  return <div></div>;
-};
-
-export { Login };
