--- src/pages/NoveltiesIndividual.js
@@ -1,6 +1,0 @@
-import React from 'react'
-import { Novelty } from '../components/Novelty/Novelty'
-
-export const NoveltiesIndividual = () => {
-    return <Novelty />
-}
