--- src/pages/Register.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-export const Register = () => {
-  return <div></div>;
-};
