+++ src/pages/backoffice/Backoffice.scss
@@ -3,8 +3,8 @@
 .Backoffice {
   position: relative;
   display: flex;
-  margin-top: 7.7rem;
   .Content {
+    border-top: 2px solid $color-black;
     width: 100%;
     min-height: 90vh; /*Quitar cuando tenga contenido */
     position: relative;
