+++ src/pages/main/About.js
@@ -1,5 +1,5 @@
 import React from "react";
 
 export const About = () => {
-  return <div></div>;
+  return <div>Hola</div>;
 };
