+++ src/pages/main/Activities.js
