+++ src/pages/main/NoveltiesIndividual.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { Novelty } from '../../components/Novelty/Novelty';
+
+export const NoveltiesIndividual = () => {
+    return <Novelty />
+}
